requirements.txt
# Activity Reporting System - Python Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # Excel file support

# PDF Generation
reportlab>=4.0.0

# API and HTTP
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Configuration
PyYAML>=6.0

# Scheduling
APScheduler>=3.10.0

# Email (built-in smtplib is used, no extra package needed)

# Optional: Data validation
# pydantic>=2.0.0

# Development and Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
